requirements.txt
numpy
scipy
google-api-python-client
google-auth-oauthlib
openai
requests
Pillow
anthropic